api/requirements.txt
alembic==1.4.2
Babel==2.8.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
dateparser==0.7.4
dialogflow==0.4.0
Flask==1.0.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
google-api-core==1.16.0
google-auth==1.11.3
googleapis-common-protos==1.51.0
grpcio==1.27.2
gunicorn==20.0.4
idna==2.6
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
protobuf==3.11.3
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-dotenv==0.8.2
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3.1
regex==2020.2.20
requests==2.18.4
rsa==4.0
six==1.14.0
SQLAlchemy==1.3.15
tzlocal==2.0.0
urllib3==1.22
Werkzeug==1.0.0
